requirements.txt
flask           # Server framework
cx-Oracle       # Interface for Oracle Database
bcrypt          # For verifying user passwords
cryptography    # For generating secure keys
oci             # Interface for Oracle Cloud Infrastrcture Object Storage (buckets)
pyjwt           # For making JWT cookies to authorize users
flask_cors      # CORS access policy for Flask (to allow requests from anywhere)
pdf2image       # For converting PDFs to PNGs
